requirements.txt
numpy
pandas
matplotlib
fpdf2
yfinance
PyPDF2
requests
